backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
PyPDF2==3.0.1
google-generativeai==0.3.0
weasyprint==59.0